fms/src/requirements.txt
flask
flask-jwt-extended
minio
python-keycloak